Rename as FullMessageIdentity; move package
